move volume detection logic to chapter utils
