+++ app/src/main/java/eu/kanade/tachiyomi/ui/reader/ReaderPresenter.kt
@@ -58,7 +58,6 @@ class ReaderPresenter(
     private val chapterFilter: ChapterFilter = Injekt.get()
 ) : BasePresenter<ReaderActivity>() {
 
-
     /**
      * The manga loaded in the reader. It can be null when instantiated for a short time.
      */
